setting-page: Refactor code for setting-page + update UI.
WIP #340.
