+++ app/renderer/css/preference.css
@@ -144,13 +144,12 @@ td:nth-child(odd) {
 }
 
 #new-server-container {
-    /* opacity: 1;
-    transition: opacity 0.3s; */
-    padding-left: 50px;
+    padding-left: 42px;
+    padding-top: 25px;
+    margin-right: 16px;
 }
 
 .title {
-    padding: 4px 0 6px 0;
     font-weight: 500;
     color: #222c31;
 }
@@ -206,13 +205,12 @@ img.server-info-icon {
 .setting-input-value {
     flex-grow: 1;
     font-size: 14px;
-    height: 22px;
     border-radius: 3px;
-    padding: 7px;
+    padding: 13px;
     border: #ededed 2px solid;
     outline-width: 0;
     background: transparent;
-    max-width: 500px;
+    max-width: 450px;
 }
 
 .setting-input-value:focus {
@@ -435,7 +433,7 @@ input.toggle-round:checked+label:after {
     display: block;
     position: fixed;
     z-index: 1;
-    padding-top: 20vh;
+    padding-top: 15vh;
     left: 0;
     top: 0;
     margin: auto;
@@ -448,12 +446,12 @@ input.toggle-round:checked+label:after {
 /* Modal Content */
 
 .modal-container {
-    background-color: #fefefe;
+    background-color: #f4f7f8;
     margin: auto;
     padding: 57px;
     border: #dae1e3 1px solid;
-    width: 600px;
-    height: 344px;
+    width: 550px;
+    height: 370px;
     border-radius: 4px;
 }
 
@@ -463,17 +461,16 @@ input.toggle-round:checked+label:after {
 }
 
 .divider {
-    width: 335px;
-    text-align: center;
-    margin-left: 0px;
-    margin-top: 48px;
+    margin-bottom: 30px;
+    margin-right: 10px;
+    margin-top: 30px;
     color: #7d878a;
 }
 
 .divider hr {
     margin-left: auto;
     margin-right: auto;
-    width: 42%;
+    width: 45%;
 }
 
 .left {
@@ -484,6 +481,23 @@ input.toggle-round:checked+label:after {
     float: right;
 }
 
+.server-center {
+    width: 100%;
+    text-align: center;
+    align-items: center;
+    padding-top: 13px;
+}
+
+.server-center span {
+    padding: 10px;
+    margin: auto;
+    align-items: center;
+    text-align: center;
+}
+
+.server-center .blue {
+    border-radius: 3px;
+}
 
 /* responsive grid */
